move button to summary tab
