+++ frontend/src/OrganizationDetails.js
@@ -1,32 +1,19 @@
 import React from 'react'
-import { useQuery, useMutation } from '@apollo/client'
-import { t, Trans } from '@lingui/macro'
-import { useLingui } from '@lingui/react'
+import { useQuery } from '@apollo/client'
+import { Trans } from '@lingui/macro'
 import {
   Box,
-  Button,
-  Flex,
   Heading,
   IconButton,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalFooter,
-  ModalHeader,
-  ModalOverlay,
   Stack,
   Tab,
   TabList,
   TabPanel,
   TabPanels,
   Tabs,
-  useDisclosure,
-  useToast,
 } from '@chakra-ui/react'
 import { ArrowLeftIcon, CheckCircleIcon } from '@chakra-ui/icons'
 import { ORG_DETAILS_PAGE } from './graphql/queries'
-import { LEAVE_ORG } from './graphql/mutations'
 import { Link as RouteLink, useParams } from 'react-router-dom'
 import { OrganizationSummary } from './OrganizationSummary'
 import { ErrorBoundary } from 'react-error-boundary'
@@ -38,8 +25,6 @@ import { useDocumentTitle } from './useDocumentTitle'
 
 export default function OrganizationDetails() {
   const { orgSlug } = useParams()
-  const toast = useToast()
-  const { i18n } = useLingui()
 
   useDocumentTitle(`${orgSlug}`)
 
@@ -47,57 +32,6 @@ export default function OrganizationDetails() {
     variables: { slug: orgSlug },
     errorPolicy: 'ignore', // allow partial success
   })
-
-  const [leaveOrg] = useMutation(LEAVE_ORG, {
-    onError(error) {
-      toast({
-        title: i18n._(t`An error occurred.`),
-        description: error.message,
-        status: 'error',
-        duration: 9000,
-        isClosable: true,
-        position: 'top-left',
-      })
-    },
-    onCompleted({ leaveOrg }) {
-      if (leaveOrg.result.__typename === 'Domain') {
-        toast({
-          title: i18n._(t`Organization left successfully`),
-          description: i18n._(t`${leaveOrg.result.domain} left ${orgSlug}`),
-          status: 'success',
-          duration: 9000,
-          isClosable: true,
-          position: 'top-left',
-        })
-        leaveOrgOnClose()
-      } else if (leaveOrg.result.__typename === 'AffiliationError') {
-        toast({
-          title: i18n._(t`Unable to leave organization.`),
-          description: leaveOrg.result.description,
-          status: 'error',
-          duration: 9000,
-          isClosable: true,
-          position: 'top-left',
-        })
-      } else {
-        toast({
-          title: i18n._(t`Incorrect send method received.`),
-          description: i18n._(t`Incorrect leaveOrg.result typename.`),
-          status: 'error',
-          duration: 9000,
-          isClosable: true,
-          position: 'top-left',
-        })
-        console.log('Incorrect leaveOrg.result typename.')
-      }
-    },
-  })
-
-  const {
-    isOpen: leaveOrgIsOpen,
-    onOpen: leaveOrgOnOpen,
-    onClose: leaveOrgOnClose,
-  } = useDisclosure()
 
   if (loading) {
     return (
@@ -111,17 +45,6 @@ export default function OrganizationDetails() {
 
   return (
     <Box w="100%" px={4}>
-      <Flex w="100%">
-        <Button
-          variant="danger"
-          onClick={() => {
-            leaveOrgOnOpen()
-          }}
-          ml="auto"
-        >
-          <Trans> Leave Organization </Trans>
-        </Button>
-      </Flex>
       <Stack isInline align="center" mb="4">
         <IconButton
           icon={<ArrowLeftIcon />}
@@ -162,6 +85,9 @@ export default function OrganizationDetails() {
                 userCount={data?.organization?.affiliations?.totalCount}
                 city={data?.organization?.city}
                 province={data?.organization?.province}
+                slug={orgSlug}
+                name={orgName}
+                id={data?.organization?.id}
               />
             </ErrorBoundary>
           </TabPanel>
@@ -179,46 +105,6 @@ export default function OrganizationDetails() {
           )}
         </TabPanels>
       </Tabs>
-
-      <Modal
-        isOpen={leaveOrgIsOpen}
-        onClose={leaveOrgOnClose}
-        motionPreset="slideInBottom"
-      >
-        <ModalOverlay />
-        <ModalContent pb={4}>
-          <ModalHeader>
-            <Trans>Leave Organization</Trans>
-          </ModalHeader>
-          <ModalCloseButton />
-          <ModalBody>
-            <Trans>
-              Are you sure you wish to leave {orgName}? You will have to be
-              invited back in to access it.
-            </Trans>
-          </ModalBody>
-
-          <ModalFooter>
-            <Button variant="primaryOutline" mr="4" onClick={leaveOrgOnClose}>
-              <Trans>Cancel</Trans>
-            </Button>
-
-            <Button
-              variant="primary"
-              mr="4"
-              onClick={async () => {
-                await leaveOrg({
-                  variables: {
-                    orgId: data?.organization?.id,
-                  },
-                })
-              }}
-            >
-              <Trans>Confirm</Trans>
-            </Button>
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
     </Box>
   )
 }
